Problem 0009: add test for dummy implementation

+++ Numbers/PythagoreanTriplets.cs
@@ -2,5 +2,5 @@
 
 public static class PythagoreanTriplets
 {
-    public static (long a, long b, long c) CreateTripledWithSum(int sum) => (3, 4, 5);
+    public static (long a, long b, long c) CreateTripledWithSum(long sum) => (3, 4, 5);
 }
